Use inverse binding adapter to save vals

+++ app/src/main/java/de/ka/jamit/tcalc/repo/Repository.kt
@@ -18,7 +18,7 @@ interface Repository {
 
     fun saveRecord(newRecord: RecordDao)
 
-    fun updateRecord(value: Int, key: Long)
+    fun updateRecord(value: Float, key: Long)
 
 
 }
